perf: opt code
